refactor(store-core): extract message/tool/stream handlers into handlers/stream.ts (audit P2-3)
Tenth slice of the P2-3 split — the largest and most complex family, the
live-chat wire path. Move the message / tool / stream rendering handlers
into a new handlers/stream.ts:

- handleMessage (+ MessagePayload)
- handleToolStart / handleToolResult / handleToolInputDelta
- handleStreamStart / handleStreamEnd
- sharedStreamDelta (+ PendingDelta / StreamDeltaContext)
- handleResultUsage / handleRawOutput

Pure move, no logic change: the moved block is byte-identical to the
origin/main block (proven by diff). stream.ts imports ActiveTool /
ChatMessage / ContextUsage / ToolResultImage (../types), nextMessageId
(../utils), isReplayDuplicate (../replay-dedup), resolveStreamId
(../stream-id), isRateLimitMessage (@chroxy/protocol) — no ./_shared (the
family resolves session ids inline). ContextUsage / ToolResultImage /
isReplayDuplicate / resolveStreamId / isRateLimitMessage were used only by
this family, so they are dropped from index.ts's imports. The barrel
re-exports via `export * from './stream'`, so the public surface is
unchanged.

handlers/index.ts: 4617 -> 3377 lines (-1240).
